Optional per-window init_cmd to run before init_cmd
Right now init_cmd is sent to all panes of all windows before sending the actual commands for that window. This commit enables an additional init_cmd_<window_name> that will run before init_cmd in all panes of the window <window_name>. This helps to set new value(s) to some variable(s) instead of using their default value(s) in that specific window.

An example test case is in window_init.yaml, where two windows are created and in each window, the ROS_MASTER is set to use a different port.

This may be useful for multi-roscore setups using rosbridge_suite.